requirements.txt
Pillow~=10.2.0
torch~=2.1.2
torchvision~=0.16.2
diffusers~=0.25.0
accelerate~=0.25.0
tqdm~=4.66.1